requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.12.0
django-environ==0.11.2
kombu==5.3.5
amqp==5.2.0
